feat: creating toggleSwitch just for Zod
